+++ terraform/pipeline/step-function.tf
@@ -107,7 +107,6 @@ resource "aws_sfn_state_machine" "sf_pipelines" {
     flatten_cqc_ratings_job_name                  = module.flatten_cqc_ratings_job.job_name
     merge_coverage_data_job_name                  = module.merge_coverage_data_job.job_name
     validate_merge_coverage_data_job_name         = module.validate_merge_coverage_data_job.job_name
-    reconciliation_job_name                       = module.reconciliation_job.job_name
 
     # crawlers
     ascwds_crawler_name      = module.ascwds_crawler.crawler_name